hopefully fixed linting errors for utf-8 patch on windows)
